Rename the ValueObjects in the Model namespace
The Model namespace was renamed to ValueObject. All usages in the code
have been changed accordingly.

Using this term for these kind of objects is closer to our DDDish
approach than Model was.
