+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Form/Service/ContactType.php
@@ -18,7 +18,7 @@
 
 namespace Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Form\Service;
 
-use Surfnet\ServiceProviderDashboard\Domain\Model\Contact;
+use Surfnet\ServiceProviderDashboard\Domain\ValueObject\Contact;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
