+++ tests/webtests/EditServiceTest.php
@@ -21,8 +21,8 @@ namespace Surfnet\ServiceProviderDashboard\Webtests;
 use Mockery as m;
 use Surfnet\ServiceProviderDashboard\Domain\Entity\Service;
 use Surfnet\ServiceProviderDashboard\Domain\Entity\Supplier;
-use Surfnet\ServiceProviderDashboard\Domain\Model\Attribute;
-use Surfnet\ServiceProviderDashboard\Domain\Model\Contact;
+use Surfnet\ServiceProviderDashboard\Domain\ValueObject\Attribute;
+use Surfnet\ServiceProviderDashboard\Domain\ValueObject\Contact;
 use Surfnet\ServiceProviderDashboard\Domain\Repository\ServiceRepository;
 use Surfnet\ServiceProviderDashboard\Domain\Repository\SupplierRepository;
 use Symfony\Bundle\FrameworkBundle\Client;
